Deduplicate calibration functions.
When looking for what function to port to C next, noticed that `SetFrequencySX` had two implementations, was like, wait, that's curious. One rabbit hole later and turns out the whole calibration stack was essentially duplicated and existed in both the LMS7002M class and just as free floating functions, and the use of it was inconsistent. This removes the "legacy" calibrations to have a single consistent calibration process.