Support dkms when it is not in PATH
This change allows CMake configuration to detect that dkms command is
not in the current PATH. This happens on some Linux systems which do
no add /usr/sbin to all users, even if they are part of sudo/wheel.
For example, this is default Raspbian OS onfiguration.

CMake's find_program() allows to detect dkms command when it is in
/usr/sbin but not in the current PATH.

The code block which runs dkms command is still kept, so that if the
dkms is uninstalled, CMake configuration adopts to it. It could be
removed in the future, but for now the idea was to minimize impact
this change is bringing.