Fix Tx stream start hang.
Sometimes Tx stream setup would get stuck waiting for thread to be ready, because worker thread notify_all would happen in the gap between main thread checking the variable and going to wait(). Even if the variable is atomic, it needs to be guarded by mutex if conditional variables waiting is involved.