gnuradio: add pregenerated python bindings files.
The gnuradio python bindings could be generated during cmake configuration step,
but the process might have issues with crawling through gnuradio headers in system path.
And the bindings generation process takes relatively long time, so just add the generated files for git tracking
and the files are marked to be regenerated automatically if neccessary.