remove extra semicolons to make old compilers happy
Extra (redundant) semicolons are valid in modern C++ (starting from
C++11). New versions of the compilers warn of this issue only if the
`Wextra-semi` option is explicitly enabled. But old compilers perform
this checking implicitly in the scope of the `Wpedantic` option.

So, the warning is not reported on modern hosts (e.g., Ubuntu 24.04 with
GCC v13.2 and Clang v18.1) but is reported on old hosts (e.g., Debian 11
with GCC v10.2 and Clang v11.0).

Although the extra semicolons may be useful in some cases, in this
particular case, they seem really redundant, so it is proposed to remove
them.

Test: build the project with GCC v10.2.
Signed-off-by: Gluttton <gluttton@ukr.net>