Use CalibrationFlag enum for channel configuration
The configuration field for calibration has changed from bool
to CalibrationFlag in some previous commit, but some code was
still assigning it as if it is boolean.

In places where calibration was not needed it is a simple
change to CalibrationFlag::NONE. In other cases use DCIQ
calibration as it seems to be the one which was actually
performed in those cases prior to the change of the configuration.