for static build export kissfft as a private lib
When statically linking to liblimesuiteng one
also needs to link against the kissft lib.
Since the user may or may not have installed
kissfft as a static lib, we provide a private
version which is linked as liblimesuiteng-kissfft-float.a
when linking against liblimesuiteng.
When the user also has installed the upstream
kissfft she may link against this version, but
then the results become dependent on where the
both libraries appear on the linker line.
Usually the first wins. So take care!